ingestion-service/requirements.txt
chromadb
requests
markdown
bs4
transformers
langchain
transformers
pydantic
fastapi
torch==2.6.0+cpu
uvicorn[standard]
redis
pdfplumber
numpy<2
sentence-transformers
qdrant-client[http] >= 1.6.0
python-dotenv
einops
openai